CivicGuard-AI/ai-verification-engine/requirements.txt
# CivicGuard-AI Verification Engine Dependencies

# Web Framework
flask==3.0.3
gunicorn==22.0.0
werkzeug==3.0.3

# Image Processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# ML/AI
scikit-learn==1.5.1
joblib==1.4.2

# Utilities
python-dateutil==2.9.0
requests==2.32.3
